+++ app/src/main/java/com/umc/product/di/ApiModule.kt
@@ -42,4 +42,10 @@ object ApiModule {
     fun provideScheduleApi(@AuthRetrofit retrofit: Retrofit): ScheduleApi {
         return retrofit.create(ScheduleApi::class.java)
     }
+
+    @Singleton
+    @Provides
+    fun provideAttendanceApi(@AuthRetrofit retrofit: Retrofit): AttendanceApi {
+        return retrofit.create(AttendanceApi::class.java)
+    }
 }
