+++ app/src/main/java/com/umc/product/di/UseCaseModule.kt
@@ -3,17 +3,10 @@ package com.umc.product.di
 import com.umc.domain.repository.AuthRepository
 import com.umc.domain.repository.member.MemberRepository
 import com.umc.domain.repository.ChallengerRepository
-import com.umc.domain.repository.schedule.ScheduleRepository
 import com.umc.domain.usecase.PostLoginUseCase
-import com.umc.domain.usecase.Schedule.GetScheduleDetailUseCase
-import com.umc.domain.usecase.Schedule.GetScheduleListUseCase
-import com.umc.domain.usecase.Schedule.GetScheduleMonthUseCase
-import com.umc.domain.usecase.appDataStore.UpdateUserInfoUseCase
 import com.umc.domain.usecase.challenger.GetChallengerDetailUseCase
-import com.umc.domain.usecase.PostLoginUseCase
 import com.umc.domain.usecase.member.GetMemberProfileUseCase
 import com.umc.domain.usecase.member.GetMyProfileUseCase
-import com.umc.domain.usecase.schedule.GetScheduleDetailUseCase
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -51,6 +44,7 @@ object UseCaseModule {
 
 
     /**schedule usecase**/
+    /*
     @Singleton
     @Provides
     fun providesGetScheduleListUseCase(repository: ScheduleRepository): GetScheduleListUseCase {
@@ -63,11 +57,8 @@ object UseCaseModule {
         return GetScheduleMonthUseCase(repository)
     }
 
-    @Singleton
-    @Provides
-    fun providesGetScheduleDetailUseCase(repository: ScheduleRepository): GetScheduleDetailUseCase {
-        return GetScheduleDetailUseCase(repository)
-    }
+
+     */
 
 
     /**challenger usecase**/
@@ -77,9 +68,7 @@ object UseCaseModule {
         return GetChallengerDetailUseCase(repository)
     }
 
-    @Singleton
-    @Provides
-    fun providesGetScheduleDetailUseCase(repository: ScheduleRepository): GetScheduleDetailUseCase {
-        return GetScheduleDetailUseCase(repository)
-    }
+
+
+
 }
