+++ data/src/main/java/com/umc/data/api/Endpoints.kt
@@ -28,18 +28,12 @@ object Endpoints {
         const val MEMBER_PROFILE = "$MEMBER/profile/{memberId}"
     }
 
-    object Schedule{
-        const val SCHEDULE = "api/v1/schedules"
-        const val DETAIL = "$SCHEDULE/{scheduleId}"
-    }
-
     object Kakao{
         const val SEARCH_LOCATION = "v2/local/search/keyword.json"
     }
 
     object Schedule{
         const val SCHEDULE = "api/v1/schedules"
-
         const val DETAIL = "$SCHEDULE/{scheduleId}"
         const val MONTH = "$SCHEDULE/my-list"
     }
