+++ data/src/main/java/com/umc/data/response/schedule/ScheduleDetailResponse.kt
@@ -4,6 +4,7 @@ import com.google.gson.annotations.SerializedName
 import com.umc.domain.model.act.check.UserCheckAvailable
 import com.umc.domain.model.enums.CategoryType
 import com.umc.domain.model.enums.CheckAvailableStatus
+import com.umc.domain.model.home.schedule.ScheduleDetailModel
 
 data class ScheduleDetailResponse(
     @SerializedName("scheduleId") val scheduleId: Int,
@@ -33,6 +34,37 @@ data class ScheduleDetailResponse(
                 longitude = longitude ?: 0.0,
                 address = locationName ?: "",
                 isLocationCertified = null
+            )
+        }
+
+        fun ScheduleDetailResponse.toHomeDomain(): ScheduleDetailModel {
+            // "T"를 기준으로 날짜와 시간을 분리
+            fun String.parseDateTime(): Pair<String, String> {
+                val dateTimeParts = this.split("T")
+                val date = dateTimeParts.getOrNull(0)?.replace("-", ".") ?: ""
+                val time = dateTimeParts.getOrNull(1)?.substring(0, 5) ?: ""
+                return Pair(date, time)
+            }
+
+            val (startDay, startTime) = startsAt.parseDateTime()
+            val (endDay, endTime) = endsAt.parseDateTime()
+
+            return ScheduleDetailModel(
+                scheduleId = scheduleId,
+                name = name,
+                description = description ?: "",
+                tags = tags ?: emptyList(),
+                startDay = startDay,
+                startTime = startTime,
+                endDay = endDay,
+                endTime = endTime,
+                isAllDay = isAllDay,
+                locationName = locationName ?: "",
+                latitude = latitude?: 0.0,
+                longitude = longitude?: 0.0,
+                status = status ?: "",
+                dDay = dDay ?: -1,
+                requiresAttendanceApproval = requiresAttendanceApproval ?: false
             )
         }
     }
